requirements.txt
beautifulsoup4==4.12.2
ctranslate2==3.20.0
Flask==2.3.3
langchain==0.0.304
numpy==1.26.0
openai==0.28.0
python-dotenv==1.0.0
Requests==2.31.0
scikit_learn==1.3.1
tqdm==4.66.1
transformers==4.33.2
